gir: generator: Generate no-arg constructors
This patch adds the no-arg constructor implementation to the generator.

Some classes (like Gtk.ComboBox) have multiple no-arg constructors and
generating both of them will result in conflicting overloads.

In case this happens, the constructor with the shortest native method
name is picked. This heuristic was chosen because most constructors use
the form `_new()` for the default constructor and
`_new_with_something()` for constructors with additional behavior.